Refactor Reportings vessel sidebar tab
