+++ frontend/src/features/VesselSidebar/Reportings/current/ConfirmDeletionModal.tsx
