+++ frontend/src/features/VesselSidebar/Reportings/current/CreateOrEditReporting.tsx
