+++ frontend/src/features/VesselSidebar/Reportings/current/utils.ts
